Rename `jl4-lsp-client` to `jl4-client-rpc`; update the package description
