+++ ts-apps/jl4-web/package.json
@@ -40,7 +40,7 @@
   },
   "dependencies": {
     "@repo/viz-expr": "*",
-    "jl4-lsp-client": "*",
+    "jl4-client-rpc": "*",
     "@codingame/monaco-vscode-api": "^14.0.4",
     "@repo/decision-logic-visualizer": "*",
     "@sveltejs/adapter-static": "^3.0.8",
